requirements.txt
msgpack-python
celery[redis]
# matplotlib
flask
Pillow==7.2.0